Fix/Add internal/renderer/v2/block_renderer.go (#1222, #1225)
